requirements.txt
git+https://github.com/huggingface/transformers@v4.56.0-Embedding-Gemma-preview
sentence-transformers>=5.0.0 
faiss-cpu 
torch 
paddleocr 
opencv-python 
pymupdf
# New dependencies for PS-05
ultralytics>=8.0.0
fasttext>=0.9.2
PyYAML>=6.0
Pillow>=9.0.0
tqdm>=4.64.0
# Training and evaluation dependencies
pycocotools>=2.0.0
jiwer>=3.0.0
nltk>=3.8.0
bert-score>=0.3.0
matplotlib>=3.5.0
markdown>=3.4.0